--- test-admin/vite.config.ts
@@ -1,14 +1,0 @@
-import { defineConfig } from 'vite';
-import react from '@vitejs/plugin-react';
-
-// https://vitejs.dev/config/
-export default defineConfig({
-    plugins: [react()],
-    define: {
-        'process.env': process.env,
-    },
-    server: {
-        host: true,
-    },
-    base: './',
-});
